+++ hdf5/src/hl/dataset.rs
@@ -1,3 +1,5 @@
+//! Interfaces for `Dataset` objects.
+
 use std::fmt::{self, Debug};
 use std::ops::Deref;
 
@@ -8,6 +10,8 @@ use hdf5_sys::h5d::{
     H5Dcreate2, H5Dcreate_anon, H5Dget_access_plist, H5Dget_create_plist, H5Dget_offset,
     H5Dset_extent,
 };
+#[cfg(feature = "1.10.0")]
+use hdf5_sys::h5d::{H5Dflush, H5Drefresh};
 use hdf5_sys::h5l::H5Ldelete;
 use hdf5_sys::h5p::H5P_DEFAULT;
 use hdf5_sys::h5z::H5Z_filter_t;
@@ -153,6 +157,22 @@ impl Dataset {
     /// Returns the pipeline of filters used in this dataset.
     pub fn filters(&self) -> Vec<Filter> {
         self.dcpl().map_or(Vec::default(), |pl| pl.filters())
+    }
+
+    /// Flush the dataset metadata from the metadata cache to the file
+    #[cfg(feature = "1.10.0")]
+    pub fn flush(&self) -> Result<()> {
+        let id = self.id();
+        h5call!(H5Dflush(id))?;
+        Ok(())
+    }
+
+    /// Refresh metadata items assosicated with the dataset
+    #[cfg(feature = "1.10.0")]
+    pub fn refresh(&self) -> Result<()> {
+        let id = self.id();
+        h5call!(H5Drefresh(id))?;
+        Ok(())
     }
 }
 
@@ -311,11 +331,15 @@ where
     }
 }
 
+/// Options for how to chunk data.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum Chunk {
-    Exact(Vec<Ix>), // exact chunk shape
-    MinKB(usize),   // minimum chunk shape in KB
-    None,           // leave it unchunked
+    /// Exact chunk shape.
+    Exact(Vec<Ix>),
+    /// Minimum chunk shape in kilobytes.
+    MinKB(usize),
+    /// Leave the data unchunked.
+    None,
 }
 
 impl Default for Chunk {